requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.2
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
numpy==1.25.2
outcome==1.2.0
pandas==2.0.3
Pillow==10.0.0
pymongo==4.4.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0
